# Conflicts:
#	backend/controllers/JogosController.php
#	backend/runtime/debug/index.data
#	backend/runtime/logs/app.log

+++ backend/controllers/JogosController.php
@@ -8,6 +8,7 @@ use app\models\Jogos;
 use app\models\Tipojogo;
 use app\models\JogosSearch;
 use yii\web\Controller;
+use yii\web\ForbiddenHttpException;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 
@@ -35,6 +36,8 @@ class JogosController extends Controller {
      * @return mixed
      */
     public function actionIndex() {
+        if (Yii::$app->user->can('admin'))
+        {
         $searchModel = new JogosSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
@@ -42,6 +45,9 @@ class JogosController extends Controller {
                     'searchModel' => $searchModel,
                     'dataProvider' => $dataProvider,
         ]);
+        }
+        else
+            throw new ForbiddenHttpException;
     }
 
     /**
@@ -51,20 +57,39 @@ class JogosController extends Controller {
      * @throws NotFoundHttpException if the model cannot be found
      */
     public function actionView($id) {
-        $tipojogo = Tipojogo::find()->orderBy('Nome')->asArray()->all();        
 
-        return $this->render('view', [
-                    'model' => $this->findModel($id),
-                    'tipojogo' => $tipojogo,
-        ]);
+        if (Yii::$app->user->can('admin'))
+        {
+            $tipojogo = Tipojogo::find()->orderBy('Nome')->asArray()->all();
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+                'tipojogo' => $tipojogo,
+            ]);
+        }
+        else
+            throw new ForbiddenHttpException;
     }
 
-    /**
+
+        /**
      * Creates a new Jogos model.
      * If creation is successful, the browser will be redirected to the 'view' page.
      * @return mixed
      */
     public function actionCreate() {
+        if (Yii::$app->user->can('admin'))
+        {
+            $model = new Jogos();
+            $tipojogo = Tipojogo::find()->orderBy('Nome')->asArray()->all();
+
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->Id]);
+            }
+
+            return $this->render('create', [
+                'model' => $model,
+                'tipojogo' => $tipojogo,
+            ]);
         $model = new Jogos();
         $tipojogo = Tipojogo::find()->orderBy('Nome')->asArray()->all();
         $genero = 2;
@@ -72,11 +97,8 @@ class JogosController extends Controller {
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
             return $this->redirect(['view', 'id' => $model->Id]);
         }
-
-        return $this->render('create', [
-                    'model' => $model,
-                    'tipojogo' => $tipojogo,
-        ]);
+        else
+            throw new ForbiddenHttpException;
     }
 
     /**
@@ -87,17 +109,22 @@ class JogosController extends Controller {
      * @throws NotFoundHttpException if the model cannot be found
      */
     public function actionUpdate($id) {
-        $model = $this->findModel($id);
-        $tipojogo = Tipojogo::find()->all();
+        if (Yii::$app->user->can('admin'))
+        {
+            $model = $this->findModel($id);
+            $tipojogo = Tipojogo::find()->all();
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->Id]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->Id]);
+            }
+
+            return $this->render('update', [
+                'model' => $model,
+                'tipojogo' => $tipojogo,
+            ]);
         }
-
-        return $this->render('update', [
-                    'model' => $model,
-                    'tipojogo' => $tipojogo,
-        ]);
+        else
+            throw new ForbiddenHttpException;
     }
 
     /**
@@ -108,9 +135,15 @@ class JogosController extends Controller {
      * @throws NotFoundHttpException if the model cannot be found
      */
     public function actionDelete($id) {
+        if (Yii::$app->user->can('admin'))
+        {
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
+        }
+        else
+            throw new ForbiddenHttpException;
+
     }
 
     /**
